Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
 int main(){
-    int a,b;
+    float a,b;
     float c;
-    scanf("%d %d",&a,&b);
+    scanf("%f %f",&a,&b);
     c=a/b;
+    
     if(c==b){
         printf("Yes");
